refactor: extract log from core
