+++ src/repl.ts
@@ -16,7 +16,7 @@ import os from 'node:os'
 import path from 'node:path'
 import { inspect } from 'node:util'
 import { ProcessOutput, defaults } from './core.js'
-import { chalk } from './vendor.js'
+import { chalk } from './vendor-core.js'
 
 export async function startRepl() {
   defaults.verbose = false
