+++ test/package.test.js
@@ -62,9 +62,6 @@ describe('package', () => {
         'build/index.cjs',
         'build/index.d.ts',
         'build/index.js',
-        'build/repl.cjs',
-        'build/repl.d.ts',
-        'build/repl.js',
         'build/util.cjs',
         'build/util.d.ts',
         'build/util.js',
